Move some common functionality into a subclass
